+++ src/HasDynamicAttributes.php
@@ -137,7 +137,7 @@ trait HasDynamicAttributes
         if ($this->dynamicDocument->has($key)) {
             $value = $this->dynamic($key);
 
-            if (is_array($value) && count(array_intersect($this->getDynamicLocales(), array_keys($value))) > 0 && in_array($locale, $this->getDynamicLocales())) {
+            if (is_array($value) && count(array_intersect($this->getDynamicLocales(), array_keys($value))) > 0) {
                 return $this->getLocalizedValue($key, $locale);
             }
 
@@ -195,16 +195,23 @@ trait HasDynamicAttributes
             $value = $this->dynamic("$key.{$locale}");
 
             // If fallback locale is given, we avoid returning null values and instead try to retrieve value via the fallback locale.
-            if (! $fallbackLocale || ! is_null($value)) {
+            if (! is_null($value)) {
                 return $value;
             }
         }
 
-        if ($this->dynamicDocument->has("$key.{$fallbackLocale}")) {
-            return $this->dynamic("$key.{$fallbackLocale}");
+        if($fallbackLocale) {
+            return $this->getLocalizedValue($key, $fallbackLocale);
         }
 
         return null;
+
+
+//        if ($this->dynamicDocument->has("$key.{$fallbackLocale}")) {
+//            return $this->dynamic("$key.{$fallbackLocale}");
+//        }
+//
+//        return null;
     }
 
     /* Override Eloquent method as part of the custom cast */