+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Changed: when returning a localized value for a locale that is not set as dynamic locale, null is returned. In previous releases, the raw array was returned instead. This often lead to render issues where a string instead of the raw array was expected.
+
 - Added: You can set locales at runtime via the `setDynamicLocales()` method.
 - Added: You can set fallback locales map via the `setDynamicFallbackLocales()` method. This expects an array with the locale as key and its fallback locale as value.
 - Added: You can set the active dynamic locale via the `setActiveDynamicLocale()` method. By default the app locale is used.